requirements.txt
moviepy>=2.0.0
pillow>=10.0.0
openai>=1.0.0
requests>=2.31.0
python-dotenv>=1.0.0
click>=8.1.0

# YouTube upload (for auto command)
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.2.0
